+++ torch/_inductor/codegen/triton.py
@@ -34,7 +34,7 @@ import torch.utils._pytree as pytree
 from torch._dynamo.utils import preserve_rng_state
 
 from torch._inductor.metrics import is_metric_table_enabled, log_kernel_metadata
-from torch._inductor.runtime.triton_heuristics import AutotuneHint
+from torch._inductor.runtime.hints import AutotuneHint
 from torch._prims_common import is_integer_dtype
 from torch.utils._sympy.functions import FloorDiv, ModularIndexing
 from torch.utils._sympy.value_ranges import ValueRanges
@@ -121,16 +121,9 @@ def gen_common_triton_imports():
 
     imports.splice(
         """
-        from torch._inductor.runtime import (
-            AutotuneHint,
-            instance_descriptor,
-            libdevice,
-            ReductionHint,
-            TileHint,
-            tl_math,
-            triton_helpers,
-            triton_heuristics,
-        )
+        from torch._inductor.runtime import triton_helpers, triton_heuristics
+        from torch._inductor.runtime.triton_helpers import libdevice, math as tl_math
+        from torch._inductor.runtime.hints import AutotuneHint, ReductionHint, TileHint, instance_descriptor
         """
     )
     return imports.getvalue()
