+++ torch/_inductor/codegen/triton_utils.py
@@ -5,7 +5,7 @@ import sympy
 import torch
 
 from .. import config
-from ..runtime import instance_descriptor
+from ..runtime.hints import instance_descriptor
 from ..utils import _type_of
 from ..virtualized import V
 from .common import KernelArgType, SizeArg, TensorArg, WorkspaceArg
