--- torch/_inductor/runtime/__init__.py
@@ -1,16 +1,0 @@
-from . import triton_helpers, triton_heuristics
-from .hints import instance_descriptor, ReductionHint, TileHint
-from .triton_helpers import libdevice, math as tl_math
-from .triton_heuristics import AutotuneHint
-
-
-__all__ = [
-    "triton_heuristics",
-    "triton_helpers",
-    "libdevice",
-    "tl_math",
-    "AutotuneHint",
-    "TileHint",
-    "ReductionHint",
-    "instance_descriptor",
-]
