+++ torch/_inductor/runtime/triton_heuristics.py
@@ -12,7 +12,6 @@ import os.path
 import re
 import threading
 import time
-from enum import auto, Enum
 from typing import Any, Callable, Dict, List, Optional, Set, Tuple
 
 import torch
@@ -24,8 +23,6 @@ from torch._dynamo.utils import dynamo_timed, get_first_attr
 from torch._inductor import config
 from torch._inductor.codecache import cache_dir, CudaKernelParamCache
 from torch._inductor.coordinate_descent_tuner import CoordescTuner
-
-from torch._inductor.runtime.hints import ReductionHint, TileHint
 from torch._inductor.utils import (
     ceildiv,
     conditional_product,
@@ -37,6 +34,13 @@ from torch._inductor.utils import (
     triton_config_to_hashable,
 )
 from torch.utils._triton import has_triton_package
+from .hints import (
+    _NUM_THREADS_PER_WARP,
+    AutotuneHint,
+    HeuristicType,
+    ReductionHint,
+    TileHint,
+)
 
 
 log = logging.getLogger(__name__)
@@ -57,28 +61,6 @@ else:
     KernelInterface = object
     OutOfResources = object
     ASTSource = None
-
-
-_NUM_THREADS_PER_WARP = 32
-
-
-class HeuristicType(Enum):
-    PERSISTENT_REDUCTION = auto()
-    POINTWISE = auto()
-    REDUCTION = auto()
-    SPLIT_SCAN = auto()
-    TEMPLATE = auto()
-    USER_AUTOTUNE = auto()
-
-
-class AutotuneHint(Enum):
-    ELEMENTS_PER_WARP_32 = 0
-
-    # Triton codegen tries to codegen set of AutotuneHints.
-    # Enum.__repr__ looks like "<AutotuneHint.ELEMENTS_PER_WARP_32: 0>""
-    # which isn't valid python.
-    # Enum.__str__ will just return "AutotuneHint.ELEMENTS_PER_WARP_32".
-    __repr__ = Enum.__str__
 
 
 def autotune_hints_to_configs(
